# Speedup C encoder by factor of 40