--- a/C/encode.c
+++ b/C/encode.c
@@ -3,7 +3,9 @@
 
 #include <string.h>
 
-static float *multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow);
+static float *multiplyBasisFunction(
+	int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow,
+	float *cosX);
 static char *encode_int(int value, int length, char *destination);
 
 static int encodeDC(float r, float g, float b);
@@ -32,14 +34,17 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 
 	init_sRGBToLinear_cache();
 
+	float *cosX = (float *)malloc(sizeof(float) * width);
+	if (! cosX) return NULL;
 	for(int y = 0; y < yComponents; y++) {
 		for(int x = 0; x < xComponents; x++) {
-			float *factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow);
+			float *factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow, cosX);
 			factors[y][x][0] = factor[0];
 			factors[y][x][1] = factor[1];
 			factors[y][x][2] = factor[2];
 		}
 	}
+	free(cosX);
 
 	float *dc = factors[0][0];
 	float *ac = dc + 3;
@@ -75,16 +80,25 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 	return buffer;
 }
 
-static float *multiplyBasisFunction(int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow) {
+static float *multiplyBasisFunction(
+	int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow,
+	float *cosX
+) {
 	float r = 0, g = 0, b = 0;
 	float normalisation = (xComponent == 0 && yComponent == 0) ? 1 : 2;
 
+	for(int x = 0; x < width; x++) {
+		cosX[x] = cosf(M_PI * xComponent * x / width);
+	}
+
 	for(int y = 0; y < height; y++) {
+		uint8_t *src = rgb + y * bytesPerRow;
+		float cosY = cosf(M_PI * yComponent * y / height);
 		for(int x = 0; x < width; x++) {
-			float basis = cosf(M_PI * xComponent * x / width) * cosf(M_PI * yComponent * y / height);
-			r += basis * sRGBToLinear_cache[rgb[3 * x + 0 + y * bytesPerRow]];
-			g += basis * sRGBToLinear_cache[rgb[3 * x + 1 + y * bytesPerRow]];
-			b += basis * sRGBToLinear_cache[rgb[3 * x + 2 + y * bytesPerRow]];
+			float basis = cosX[x] * cosY;
+			r += basis * sRGBToLinear_cache[src[3 * x + 0]];
+			g += basis * sRGBToLinear_cache[src[3 * x + 1]];
+			b += basis * sRGBToLinear_cache[src[3 * x + 2]];
 		}
 	}