--- a/C/encode.c
+++ b/C/encode.c
@@ -5,7 +5,7 @@
 
 static float *multiplyBasisFunction(
 	int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow,
-	float *cosX);
+	float *cosX, float *cosY);
 static char *encode_int(int value, int length, char *destination);
 
 static int encodeDC(float r, float g, float b);
@@ -34,17 +34,32 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 
 	init_sRGBToLinear_cache();
 
-	float *cosX = (float *)malloc(sizeof(float) * width);
+	float *cosX = (float *)malloc(sizeof(float) * width * xComponents);
 	if (! cosX) return NULL;
+	float *cosY = (float *)malloc(sizeof(float) * height);
+	if (! cosY) {
+		free(cosX);
+		return NULL;
+	}
+	for(int x = 0; x < xComponents; x++) {
+		for(int i = 0; i < width; i++) {
+			cosX[x * width + i] = cosf(M_PI * x * i / width);
+		}
+	}
 	for(int y = 0; y < yComponents; y++) {
+		for(int i = 0; i < height; i++) {
+			cosY[i] = cosf(M_PI * y * i / height);
+		}
 		for(int x = 0; x < xComponents; x++) {
-			float *factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow, cosX);
+			float *factor = multiplyBasisFunction(x, y, width, height, rgb, bytesPerRow,
+				cosX + x * width, cosY);
 			factors[y][x][0] = factor[0];
 			factors[y][x][1] = factor[1];
 			factors[y][x][2] = factor[2];
 		}
 	}
 	free(cosX);
+	free(cosY);
 
 	float *dc = factors[0][0];
 	float *ac = dc + 3;
@@ -82,20 +97,15 @@ const char *blurHashForPixels(int xComponents, int yComponents, int width, int h
 
 static float *multiplyBasisFunction(
 	int xComponent, int yComponent, int width, int height, uint8_t *rgb, size_t bytesPerRow,
-	float *cosX
+	float *cosX, float *cosY
 ) {
 	float r = 0, g = 0, b = 0;
 	float normalisation = (xComponent == 0 && yComponent == 0) ? 1 : 2;
 
-	for(int x = 0; x < width; x++) {
-		cosX[x] = cosf(M_PI * xComponent * x / width);
-	}
-
 	for(int y = 0; y < height; y++) {
 		uint8_t *src = rgb + y * bytesPerRow;
-		float cosY = cosf(M_PI * yComponent * y / height);
 		for(int x = 0; x < width; x++) {
-			float basis = cosX[x] * cosY;
+			float basis = cosY[y] * cosX[x];
 			r += basis * sRGBToLinear_cache[src[3 * x + 0]];
 			g += basis * sRGBToLinear_cache[src[3 * x + 1]];
 			b += basis * sRGBToLinear_cache[src[3 * x + 2]];