There are benchmarks for 2000 × 1334 jpeg:

```
## Master
/Code/blurhash/c# make && ./blurhash_encoder 6 4 /Code/imgs/bologna2k.jpg 
Time per 3 execution: 2623.813 ms
W7E-z7oyM{8xM{wKwdMepHrE%LV[OVV@BBS$r@NaR7OrRQNaMKXm

/Code/blurhash/c# make && ./blurhash_encoder 9 9 /Code/imgs/bologna2k.jpg 
Time per 3 execution: 9159.889 ms
|7E-z7oyM{8xM{wKTdWFNuwdMepHrE%LV[kVV@jFOVV@BBS$r@Naen={rrR7OrRQNaMKXmw|?aRPaLt6w0x[RQkqa1xuD*xZNuWEt6n%s-Vtf+M|x[ixxswctQnPNGR-RkXma0xtRkS~jF$%WXI;S~M|X8RjxYniR+ozs-

## sRGBToLinear_cache
/Code/blurhash/c# make && ./blurhash_encoder 6 4 /Code/imgs/bologna2k.jpg 
Time per 3 execution: 574.121 ms
W7E-z7oyM{8xM{wKwdMepHrE%LV[OVV@BBS$r@NaR7OrRQNaMKXm

## cosX cache
/Code/blurhash/c# make && ./blurhash_encoder 6 4 /Code/imgs/bologna2k.jpg 
Time per 10 execution: 102.816 ms
W7E-z7oyM{8xM{wKwdMepHrE%LV[OVV@BBS$r@NaR7OrRQNaMKXm

/Code/blurhash/c# make && ./blurhash_encoder 9 9 /Code/imgs/bologna2k.jpg 
Time per 10 execution: 345.734 ms
|7E-z7oyM{8xM{wKTdWFNuwdMepHrE%LV[kVV@jFOVV@BBS$r@Naen={rrR7OrRQNaMKXmw|?aRPaLt6w0x[RQkqa1xuD*xZNuWEt6n%s-Vtf+M|x[ixxswctQnPNGR-RkXma0xtRkS~jF$%WXI;S~M|X8RjxYniR+ozs-

## Single pass (final commit)
/Code/blurhash/c# make && ./blurhash_encoder 6 4 /Code/imgs/bologna2k.jpg 
Time per 10 execution: 64.382 ms
W7E-z7oyM{8xM{wKwdMepHrE%LV[OVV@BBS$r@NaR7OrRQNaMKXm

/Code/blurhash/c# make && ./blurhash_encoder 9 9 /Code/imgs/bologna2k.jpg 
Time per 10 execution: 212.955 ms
|7E-z7oyM{8xM{wKTdWFNuwdMepHrE%LV[kVV@jFOVV@BBS$r@Naen={rrR7OrRQNaMKXmw|?aRPaLt6w0x[RQkqa1xuD*xZNuWEt6n%s-Vtf+M|x[ixxswctQnPNGR-RkXma0xtRkS~jF$%WXI;S~M|X8RjxYniR+ozs-
```