@homm 2024-10-11
@DagAgren Are you interested in this improvements?